+++ src/main/java/com/hoolink/sdk/bo/omc/WorkParamBO.java
@@ -9,8 +9,8 @@ import lombok.Data;
 @Data
 public class WorkParamBO extends PageParamBO {
     private Long projectId;
-    private Integer priority;
+    private Byte priority;
     private String type;
-    private Integer orderStatus;
+    private Byte orderStatus;
     private Boolean overdue;
 }
